# src/main/resources/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:10420","description":"Generated server url"}],"paths":{"/article/{articleId}/unlike":{"put":{"tags":["article-controller"],"operationId":"unlikeArticle","parameters":[{"name":"Auth-Token","in":"header","required":true,"schema":{"type":"string"}},{"name":"articleId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}},"/article/{articleId}/like":{"put":{"tags":["article-controller"],"operationId":"likeArticle","parameters":[{"name":"Auth-Token","in":"header","required":true,"schema":{"type":"string"}},{"name":"articleId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}},"/article":{"get":{"tags":["article-controller"],"operationId":"getArticles","parameters":[{"name":"Auth-Token","in":"header","required":true,"schema":{"type":"string"}},{"name":"author","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"title","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"lowerIndex","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"upperIndex","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":-1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleSimpleListDTO"}}}}}},"put":{"tags":["article-controller"],"operationId":"updateArticle","parameters":[{"name":"Accept","in":"header","schema":{"type":"string","enum":["application/json"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleInsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdDTO"}}}}}},"post":{"tags":["article-controller"],"operationId":"insertArticle","parameters":[{"name":"Accept","in":"header","schema":{"type":"string","enum":["application/json"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleInsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdDTO"}}}}}}},"/user/register":{"post":{"tags":["user-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppUserWithPasswordDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/user/login":{"post":{"tags":["user-controller"],"operationId":"login","parameters":[{"name":"Accept","in":"header","schema":{"type":"string","enum":["application/json"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string","format":"uuid"}}}}}}},"/article/comment":{"post":{"tags":["article-controller"],"operationId":"insertComment","parameters":[{"name":"Accept","in":"header","schema":{"type":"string","enum":["application/json"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentInsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/user/{uuid}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AppUserDTO"}}}}}}},"/user/{uuid}/detailed":{"get":{"tags":["user-controller"],"operationId":"getUserDetailed","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AppUserDetailedDTO"}}}}}}},"/user/{userId}/actions/{articleId}":{"get":{"tags":["user-controller"],"operationId":"getActions","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}},{"name":"articleId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/article/{id}":{"get":{"tags":["article-controller"],"operationId":"getArticleById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"Auth-Token","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleDetailedDTO"}}}}}}},"/article/{articleId}/export":{"get":{"tags":["article-controller"],"operationId":"exportArticle","parameters":[{"name":"articleId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DataDTO"}}}}}}},"/article/{articleId}":{"delete":{"tags":["article-controller"],"operationId":"deleteArticle","parameters":[{"name":"articleId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"ArticleInsertDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"authors":{"type":"array","items":{"type":"string","format":"uuid"}},"title":{"type":"string"},"categories":{"type":"array","items":{"type":"string"}},"content":{"type":"string"}}},"IdDTO":{"type":"object","properties":{"id":{"type":"string"}}},"AppUserWithPasswordDTO":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"userName":{"type":"string"},"passwordHash":{"type":"string"}}},"LoginRequestDTO":{"type":"object","properties":{"username":{"type":"string"},"passwordHash":{"type":"string"}}},"CommentInsertDTO":{"type":"object","properties":{"content":{"type":"string"},"articleId":{"type":"string","format":"uuid"},"appUserId":{"type":"string","format":"uuid"}}},"AppUserDTO":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"userName":{"type":"string"}}},"AppUserDetailedDTO":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"userName":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"roles":{"type":"array","items":{"type":"string"}}}},"ArticleSimpleDTO":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"authors":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/AppUserDTO"}},"categories":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/CategoryDTO"}},"publisher":{"$ref":"#/components/schemas/PublisherDTO"},"likeCount":{"type":"integer","format":"int32"},"liked":{"type":"boolean"}}},"ArticleSimpleListDTO":{"type":"object","properties":{"articles":{"type":"array","items":{"$ref":"#/components/schemas/ArticleSimpleDTO"}},"hasMore":{"type":"boolean"}}},"CategoryDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"}}},"PublisherDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"}}},"ArticleDetailedDTO":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"authors":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/AppUserDTO"}},"categories":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/CategoryDTO"}},"publisher":{"$ref":"#/components/schemas/PublisherDTO"},"liked":{"type":"boolean"},"likeCount":{"type":"integer","format":"int32"},"comments":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/CommentDTO"}}}},"CommentDTO":{"type":"object","properties":{"id":{"type":"string"},"content":{"type":"string"},"author":{"$ref":"#/components/schemas/AppUserDTO"}}},"DataDTO":{"type":"object","properties":{"data":{"type":"string"}}}}}}